requirements.txt
azure==4.0.0
azure-cognitiveservices-vision-computervision==0.9.0
azure-cognitiveservices-vision-customvision==3.1.0
azure-cognitiveservices-vision-face==0.6.0
azure-ai-formrecognizer==3.1.2
azure-ai-textanalytics==5.1.0
PyYAML==6.0
pillow==9.1.0
requests==2.27.1
msrest==0.6.21